Carve functional defcall* from macro defcall
defcall is defined as a macro.  It is not a light-weight container on
top of some functional definition, rather it is a macro with some custom
logic inside of it.  This form of development often occurs, yet it can
have annoying side-effects when building on top of these abstractions.
This is what occurred when trying out some plausible abstractions built
on top of this.  For this reason, we're carving out a functional variant
of defcall, which the macro may start using in the future.