Correctly interpret parsed rationals
When jsown parses a float, it may yield it as a rational.  The value is
the same, but when trying to input it into a query, things go haywire.
For the sparql query to be valid, we must express floats as a number
with a dot in it (rather than a slash).  For this reason, we now coerce
the result if the content is a float.