Removing unused imports
This library required uuid and local-time, both of which should likely
have been required in dependent libraries.  Removing in favour of
requirement in depending libraries.

It may be that some of these get added back in mu-cl-support if
mu-cl-support would choose to offer sensible defaults.  However, that
would require more libraries to be based on mu-cl-support.  We therefore
consider this to be future work.