In clean-string \ should be escaped before "
Both \ as well as " are special characters.  The quotes need to be
escaped with backslashes.  However, backslashes need to be escaped too.
This may cause some weird queries with backslashes being escaped by
multiple systems, but it aught to work :)

Why we need double escaping for the \, I'm not sure.  A \ should become
\\, but cl-ppcre seems to parse it too, hence it expects \\\\.  In order
to enter that in a lisp string, we need \\\\\\\\.  Yay for clarity /o\