Abstracting out used hunchentoot functionality
This is a step for allowing different backends to be written.  At this
stage we're preparing the code-base so it would technically be possible
to carve out this code and offer it as other services.  In its current
state this would mean that you can fork this library and conditionally
remove content, or that you supply a full alternative to mu-support with
mu-support being the basis.

Our use-case for this is running mu.semte.ch on Weblogic, which requires
a different webserver and then some.