Passing through mu-auth-allowed-groups and mu-auth-used-groups
These groups are used to identify the access rights of a user and to
help with caching.

mu-auth-used-groups is not implemented correctly yet.  It is the
responsibility of the microservice to group the used
mu-auth-used-groups together as a set of json objects.  We currently
yield the last returned response which is correct for the
current (limited) implementation of the mu-authorization service.